further cleaning up http_file mod
